Moved languages obj to index.js
